Restructure Actions decoration to be more aligned with the to be official lib support
