+++ github/actions_runner_test.go
@@ -28,7 +28,7 @@ func TestListActionRunners(t *testing.T) {
 	assert := assert.New(t)
 	org := "philips-internal"
 	repo := "fact-service"
-	runners, _, err := client.ListActionsRunners(ctx, org, repo)
+	runners, _, err := client.Actions.ListRunners(ctx, org, repo)
 	assert.NoError(err, "Failed listing action runners for %s/%s", org, repo)
 	assert.NotEmpty(runners)
 }
@@ -37,7 +37,7 @@ func TestListActionRunnerDownloads(t *testing.T) {
 	assert := assert.New(t)
 	org := "philips-internal"
 	repo := "fact-service"
-	downloads, _, err := client.ListActionsRunnersDownloads(ctx, org, repo)
+	downloads, _, err := client.Actions.ListRunnersDownloads(ctx, org, repo)
 	assert.NoError(err, "Failed listing action runner downloads for %s/%s", org, repo)
 	assert.NotEmpty(downloads)
 	assert.Len(downloads, 5, "Expected to have 5 downloads")
